requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
SQLAlchemy==2.0.40
cryptography==36.0.2
elevenlabs==1.57.0
python-dotenv==1.0.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
WTForms==3.2.1
google-generativeai
Markdown==3.8
